Fixes Gradle build, Fixes unit tests
Adds build.gradle so we can build and run unit tests.

Adds Class StringLinkedList to create a string Linked List.

Adds function “add” to add a string to the linked list.

Works on parseIn function to work with the “add” function so we can add
the string to the linked list.